Migración a PostgreSQL, catálogos en BD, cifrado de tokens e infra AWS
Base de datos:
- Migra de SQLite a PostgreSQL (asyncpg) en dev y prod; elimina SQLite y Alembic
- Tablas se crean solas al arrancar (create_all) + seed de catálogos
- Montos en Numeric(12,2); pool con pool_pre_ping para RDS

Modelo de datos configurable por BD:
- Tablas categories y bank_senders (antes hardcodeadas en código)
- Nuevo endpoint GET /categories; categoría skincare; parser de Amex
- Validación de categoría contra BD en vez de Literal

Seguridad y perfil:
- Cifra google_access/refresh_token con Fernet (EncryptedString)
- Campos monthly_income y savings_percent en User (PATCH /auth/me)

Sync de Gmail:
- Fecha de transacción desde internalDate del correo (zona MX), no la del sync
- Sync incremental desde el último correo; persiste token refrescado
- Llamada a Gmail en threadpool; fallback de categorizador ante errores de API

Infraestructura (terraform/):
- EC2 t4g.micro + RDS Postgres db.t4g.micro, us-east-1, VPC default
- HTTPS con Caddy + sslip.io (sin dominio); EIP fija; budget alerts
- Scripts deploy/stop/start para operar y ahorrar créditos

+++ app/models/budget.py
@@ -1,6 +1,7 @@
 import uuid
 from datetime import datetime, timezone
-from sqlalchemy import String, Float, DateTime, ForeignKey, UniqueConstraint
+from decimal import Decimal
+from sqlalchemy import String, Numeric, DateTime, ForeignKey, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.core.database import Base
 
@@ -12,7 +13,7 @@ class Budget(Base):
     id: Mapped[str] = mapped_column(String, primary_key=True, default=lambda: str(uuid.uuid4()))
     user_id: Mapped[str] = mapped_column(String, ForeignKey("users.id"), nullable=False, index=True)
     category: Mapped[str] = mapped_column(String, nullable=False)
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     month: Mapped[str] = mapped_column(String, nullable=False)  # formato YYYY-MM
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
     updated_at: Mapped[datetime] = mapped_column(
